qlog: rename qlog.go to connection_tracer.go
