+++ qlog/connection_tracer.go
@@ -6,7 +6,6 @@ import (
 	"io"
 	"log"
 	"net"
-	"runtime/debug"
 	"time"
 
 	"github.com/quic-go/quic-go/internal/protocol"
@@ -16,34 +15,6 @@ import (
 
 	"github.com/francoispqt/gojay"
 )
-
-// Setting of this only works when quic-go is used as a library.
-// When building a binary from this repository, the version can be set using the following go build flag:
-// -ldflags="-X github.com/quic-go/quic-go/qlog.quicGoVersion=foobar"
-var quicGoVersion = "(devel)"
-
-func init() {
-	if quicGoVersion != "(devel)" { // variable set by ldflags
-		return
-	}
-	info, ok := debug.ReadBuildInfo()
-	if !ok { // no build info available. This happens when quic-go is not used as a library.
-		return
-	}
-	for _, d := range info.Deps {
-		if d.Path == "github.com/quic-go/quic-go" {
-			quicGoVersion = d.Version
-			if d.Replace != nil {
-				if len(d.Replace.Version) > 0 {
-					quicGoVersion = d.Version
-				} else {
-					quicGoVersion += " (replaced)"
-				}
-			}
-			break
-		}
-	}
-}
 
 const eventChanSize = 50
